net: mpls: Add MPLS code
In conventional IP forwarding, a particular router will typically
consider two packets to be in the same FEC if there is some address
prefix X in that router's routing tables such that X is the "longest
match" for each packet's destination address. As the packet
traverses the network, each hop in turn reexamines the packet and
assigns it to a FEC.

In MPLS, the assignment of a particular packet to a particular FEC is
done just once, as the packet enters the MPLS tunnel. The FEC to which
the packet is assigned is encoded as a short fixed length value known
as a "label". When a packet is forwarded to its next hop, the label
is sent along with it. To be more precise, the packets are "labeled"
before they are forwarded.

At subsequent hops, there is no further analysis of the packet's
network layer header. Rather, the label is used as an index into a
table which specifies the next hop, and a new label. The old label
is replaced with the new label, and the packet is forwarded to its
next hop.

In the MPLS forwarding paradigm, once a packet is assigned to a FEC,
no further header analysis is done by subsequent routers. All
forwarding is driven by the labels.

MPLS can be built as a module, or it can be built in kernel, depending
on the selected configuration.

sysctl, proc_fs and init/exit functions are located in file

net/mpls/mpls_core.c

MPLS's sysctl and proc_fs code are fully aware of network namespaces.

All packet manipulation functions (label push, swap, pop ...), that are
necessary for MPLS to function are defined in two files:

net/mpls/mpls_nhlfe.c
net/mpls/mpls_cmd.h

IP{v4, v6} packet encapsulation in MPLS tunnels and sending them to
appropriate next hop is done via MPLS virtual devices.
Code for manipulating MPLS virtual devices is located in file:

net/mpls/mpls_dev.c

MPLS packet receiving and forwarding is done in file

net/mpls/mpls_ilm.c

Incoming Label Map (ILM) entries are stored in radix tree. ILM tree's
are manipulated per network namespace.

MPLS currently support only IP{v4, v6} {de, en}capsulation.