net: Add aditional argument to callback function from ip{,6}_fragment
Additional argument represents pointer to data, that should be forwarded to callback function.

To preserve old ip{,6}_fragment API, ip{,6}_fragment with new callback function is renamed to __ip{,6}_fragment and ip{,6}_fragment is an inline wrapper for calling __ip{,6}_fragment with old argumets.

This is needed for MPLS forwarding, so we could forward existing MPLS stack or NHLFE.