net: Large number of fixes and improvements
1. nhlfe don't hold nlattr but each instruction has it field.
At the end of the NHLFE we store MPLS labels, that are network ready.
If we have swap instruction, first label is for swaping.
Other labels are for pushing. NHLFE is ALIGNED to sizeof(long long)
2. MPLSCB now contains IPCB and IP6CB structs inside of itself. That way
we can pass sizeof MPLS stack to the ip{,6}_fragment function. Size of MPLSCB
must not be larger then the size of sk_buff member cb!
3. Fixed sending of ICMP packets and fragmentation of MPLS packets
4. Fixed error handling on the MPLS packet path (skb's were freed
twice in case of error, etc.)
5. Netns handling is tested and looks ok