net: Introduce nf_mpls struct
Struct nf_mpls is now caried inside of struct sk_buff as one of its members.

It's used the same way as nf_bridge is used. It carries MPLS private data,
as nexthop addr, sizeof MPLS stack, MPLS stack, pointers to nhlfe and to incoming net_device.

Nexthop addr is needed in the IPv{4,6} layer so the packet would be sent to the appropriate nexthop.
Sizeof MPLS stack, MPLS stack, nhlfe and dev are needed when we're fragmenting MPLS packet, or when we're
sending ICMP messages from MPLS tunnel.